+++ src/assets/basketArray.js
@@ -1,0 +1,16 @@
+export const data_basket = [
+    {
+        key: '13',
+        image: 'https://static.chipdip.ru/lib/313/DOC005313873.jpg',
+        name: '1N4148WS-7-F',
+        discription: 'Диод 150мА 75В [SOD-323]',
+        price: '40',
+    },
+    {
+        key: '14',
+        image: 'https://static.chipdip.ru/lib/464/DOC002464721.jpg',
+        name: '2А517А-2',
+        discription: 'Диод СВЧ',
+        price: '50'
+    },
+];
