+++ src/components/Tables/Tables.css
@@ -1,0 +1,29 @@
+.component__image {
+    width: 50px;
+    height: 50px;
+    margin: 0 10px;
+}
+
+.component__rate {
+    font-size: 12px;
+}
+
+.component__tag {
+    margin-bottom: 1ch;
+    font-size: 10px;
+}
+
+.component__backet-active {
+    color: #2ab234;
+}
+.component__backet-active:hover {
+    color: #78de80;
+}
+
+.component__heart-active {
+    color: #ec386e;
+}
+
+.basket-right__total {
+    margin-top: 10%;
+}
